backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.0.1
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4
channels==4.0.0
channels-redis==4.1.0
django-filter==23.3
djangorestframework-simplejwt==5.3.0
stripe==7.8.0
firebase-admin==6.2.0
gunicorn==21.2.0
whitenoise==6.6.0
boto3==1.29.7
django-storages==1.14.2
drf-spectacular==0.26.5
django-debug-toolbar==4.2.0
pytest==7.4.3
pytest-django==4.5.2
black==23.11.0
flake8==6.1.0
coverage==7.3.2
factory-boy==3.3.0
django-extensions==3.2.3